Lay the groundwork to allow for different shell profiles.
Mostly just for the fun of it, really, though I have been in a couple of situations where
I thought, "oh, it would be useful to have a custom shell for this project", and a tool
that makes it easier to develop these kinds of shells isn't an altogether bad idea.

Probably overengineering this.